Refactoring project, move everything inside src to main project structure
