+++ .travis.yml
@@ -4,7 +4,22 @@ services:
     - docker
 
 go:
-    - 1.11.x
+    - 1.12.x
+    - 1.13.x
 
-script: 
+cache:
+  directories:
+    - $GOPATH/pkg/mod
+
+env:
+  global:
+    - GO111MODULE=on
+
+install:
+  - go mod tidy
+  - git diff --exit-code go.mod
+  - git diff --exit-code go.sum
+  - go mod download
+
+script:
     - go test -v ./...
