+++ cmd/root.go
@@ -1,0 +1,35 @@
+package cmd
+
+import (
+	"os"
+	"strings"
+
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+)
+
+// rootCmd represents the base command when called without any sub-commands
+var rootCmd = &cobra.Command{
+	Use:   "bindman-dns-bind9",
+	Short: "Manages Bind9 DNS Server instances",
+	Long:  "NSUpdate commands get dispatched from REST API calls",
+}
+
+// Execute adds all child commands to the root command and sets flags appropriately.
+// This is called by main.main(). It only needs to happen once to the rootCmd.
+func Execute() {
+	if err := rootCmd.Execute(); err != nil {
+		os.Exit(1)
+	}
+}
+
+func init() {
+	cobra.OnInitialize(initConfig)
+}
+
+// initConfig reads ENV variables if set.
+func initConfig() {
+	viper.SetEnvPrefix("BINDMAN")
+	viper.SetEnvKeyReplacer(strings.NewReplacer("-", "_", ".", "_"))
+	viper.AutomaticEnv() // read in environment variables that match
+}
