+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/labbsr0x/bindman-dns-bind9/cmd"
+
+func main() {
+	cmd.Execute()
+}
