--- src/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/labbsr0x/bindman-dns-bind9/src/cmd"
-
-func main() {
-	cmd.Execute()
-}
